Attempted to remove the offset between nmrshiftdb and mnova in the atom index
Attempted to remove the offset between nmrshiftdb and mnova in the atom index

Checked the whole process and found the assign step was not working. Introduced the error when multiple datasets with the same experiment type allowed and an increment was added to the expt names ie HSQC_0. This has been corrected in simpleAssign_eeh.qs